relocate stuff from Chat.init()
