+++ src/js/components/SideBar.js
@@ -1,6 +1,7 @@
 import { html, Component } from '../lib/htm.preact.js';
 import { translate as t } from '../Translation.js';
 import {localState} from '../Main.js';
+import {showNewChat} from '../Chat.js';
 import ChatListItem from './ChatListItem.js';
 
 class SideBar extends Component {
@@ -25,6 +26,11 @@ class SideBar extends Component {
       chats[id] = chat;
       debouncedUpdate();
     });
+  }
+
+  onNewChatClick() {
+    showNewChat();
+    localState.get('activeChat').put('new');
   }
 
   render() {
@@ -59,7 +65,7 @@ class SideBar extends Component {
         </svg>
         ${t('public_messages')}
         </div>
-        <div class="chat-item new ${this.state.activeChat === 'new' ? 'active' : ''}" onClick=${() => localState.get('activeChat').put('new')}>
+        <div class="chat-item new ${this.state.activeChat === 'new' ? 'active' : ''}" onClick=${() => this.onNewChatClick()}>
           <svg class="svg-inline--fa fa-smile fa-w-16" style="margin-right:10px;margin-top:3px" version="1.1" id="Capa_1" xmlns="http://www.w3.org/2000/svg" xmlns:xlink="http://www.w3.org/1999/xlink" x="0px" y="0px"
               viewBox="0 0 510 510" xml:space="preserve">
             <path fill="currentColor" d="M459,0H51C22.95,0,0,22.95,0,51v459l102-102h357c28.05,0,51-22.95,51-51V51C510,22.95,487.05,0,459,0z M102,178.5h306v51 H102V178.5z M306,306H102v-51h204V306z M408,153H102v-51h306V153z"/>
